pyproject.toml
[tool.poetry]
name = "a-sudoku-p"
version = "0.1.0"
description = "A basic Python-DLV program for sudoku solving"
authors = ["pierluigi trocini"]
license = "My driving license is enough"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
embasp-python = "^8.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
